blog_task/requirements.txt
autopep8==1.4.4
Django==2.2.6
django-crispy-forms==1.7.2
django-environ==0.4.5
django-tinymce4-lite==1.7.5
jsmin==2.2.2
Pillow==6.2.0
psycopg2==2.8.3
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
